feat: calendar 페이지 detail slide
