+++ client/src/Components/molecule/Dropdown/style.scss
@@ -72,13 +72,3 @@
   width: 100%;
   height: 100vh;
 }
-.background {
-  position: fixed;
-  z-index: 10;
-  width: 100%;
-  height: 100vh;
-  left: 0;
-  top: 0;
-  background-color: $title-active;
-  opacity: 0.5;
-}
