+++ client/src/scss/index.scss
@@ -56,3 +56,14 @@ body {
   animation: drift 19s infinite linear;
   background: $primary1;
 }
+
+.background {
+  position: fixed;
+  z-index: 10;
+  width: 100%;
+  height: 100vh;
+  left: 0;
+  top: 0;
+  background-color: $title-active;
+  opacity: 0.5;
+}
